+++ src/main/java/com/shynieke/statues/blocks/decorative/charity/EagleRayStatueBlock.java
@@ -19,6 +19,11 @@ public class EagleRayStatueBlock extends AbstractStatueBase {
 	}
 
 	@Override
+	public boolean isHiddenStatue() {
+		return true;
+	}
+
+	@Override
 	public VoxelShape getShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
 		return SHAPE;
 	}